Revert "Move native_firm/payload/include to native_firm/include"
This reverts commit 2cf3c473dc3a9fea95cf937b467a8b75ae858868.
