+++ rxmode/native_firm/source/payload/include/vars.h
